requirements.txt
ccxt==4.4.69
tiktoken==0.8.0
aiohttp==3.10.11
aiosqlite==0.20.0
aiohttp-client-cache==0.12.4
rich==13.9.4
numpy==2.1.3
numba==0.61.0
scipy==1.15.1
pandas==2.2.3
audioop-lts==0.2.1
discord.py==2.4.0
plotly==6.0.1
markdown==3.8